backend/core/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
gunicorn